Fixed Icy talons & generalized rotation code
