+++ proto/deathknight.proto
@@ -152,8 +152,6 @@ message DeathKnight {
 		bool unholy_presence_opener = 3;
 		double disease_refresh_duration = 4;
 		bool refresh_horn_of_winter = 5;
-
-		bool wip_frost_rotation = 6;
 	}
 	Rotation rotation = 1;
