+++ sim/deathknight/rotation.go
@@ -10,13 +10,9 @@ import (
 )
 
 func (deathKnight *DeathKnight) OnAutoAttack(sim *core.Simulation, spell *core.Spell) {
-	if !deathKnight.Talents.HowlingBlast {
-		deathKnight.tryUseGCD(sim)
-	} else {
-		if !deathKnight.DKRotation.onOpener {
-			if deathKnight.GCD.IsReady(sim) {
-				deathKnight.tryUseGCD(sim)
-			}
+	if !deathKnight.onOpener {
+		if deathKnight.GCD.IsReady(sim) {
+			deathKnight.tryUseGCD(sim)
 		}
 	}
 }
@@ -43,20 +39,113 @@ func (deathKnight *DeathKnight) spreadDiseases(sim *core.Simulation, target *cor
 }
 
 func (deathKnight *DeathKnight) tryUseGCD(sim *core.Simulation) {
-	//var spell *core.Spell
-	var target = deathKnight.CurrentTarget
-
 	if deathKnight.GCD.IsReady(sim) {
-		// UH DK rota
-		if deathKnight.Talents.SummonGargoyle {
+		deathKnight.DoRotation(sim)
+	}
+}
+
+func (o *Opener) DoNext(sim *core.Simulation, deathKnight *DeathKnight) bool {
+	target := deathKnight.CurrentTarget
+	casted := &deathKnight.castSuccessful
+	advance := true
+	*casted = false
+
+	if o.idx < o.numActions {
+		action := o.actions[o.idx]
+
+		switch action {
+		case OpenerAction_IT:
+			*casted = deathKnight.CastIcyTouch(sim, target)
+			// Add this line if you care about recasting a spell in the opener in
+			// case it missed
+			advance = deathKnight.LastCastOutcome != core.OutcomeMiss
+		case OpenerAction_PS:
+			*casted = deathKnight.CastPlagueStrike(sim, target)
+			advance = deathKnight.LastCastOutcome != core.OutcomeMiss
+		case OpenerAction_UA:
+			*casted = deathKnight.CastUnbreakableArmor(sim, target)
+			// Add this line if your spell does not incur a GCD or you will hang!
+			deathKnight.WaitUntil(sim, sim.CurrentTime)
+		case OpenerAction_BT:
+			*casted = deathKnight.CastBloodTap(sim, target)
+			deathKnight.WaitUntil(sim, sim.CurrentTime)
+		case OpenerAction_Obli:
+			*casted = deathKnight.CastObliterate(sim, target)
+		case OpenerAction_FS:
+			*casted = deathKnight.CastFrostStrike(sim, target)
+		case OpenerAction_Pesti:
+			*casted = deathKnight.CastPestilence(sim, target)
+			if deathKnight.LastCastOutcome == core.OutcomeMiss {
+				advance = false
+			}
+		case OpenerAction_ERW:
+			*casted = deathKnight.CastEmpowerRuneWeapon(sim, target)
+			deathKnight.WaitUntil(sim, sim.CurrentTime)
+		case OpenerAction_HB_Ghoul_RimeCheck:
+			// You can do custom actions, this is deciding whether to HB or raise dead
+			if deathKnight.RimeAura.IsActive() {
+				*casted = deathKnight.CastHowlingBlast(sim, target)
+			} else {
+				*casted = deathKnight.CastRaiseDead(sim, target)
+			}
+		case OpenerAction_BS:
+			*casted = deathKnight.CastBloodStrike(sim, target)
+		}
+
+		// Advances the opener
+		if *casted && advance {
+			o.idx += 1
+		}
+	} else {
+		deathKnight.onOpener = false
+
+		if deathKnight.opener.id == OpenerID_FrostSubBlood_Full {
+			if deathKnight.ShouldHornOfWinter(sim) {
+				*casted = deathKnight.CastHornOfWinter(sim, target)
+			} else {
+				*casted = deathKnight.DiseaseCheckWrapper(sim, target, deathKnight.Obliterate)
+				if !*casted {
+					if deathKnight.KillingMachineAura.IsActive() && !deathKnight.RimeAura.IsActive() {
+						*casted = deathKnight.DiseaseCheckWrapper(sim, target, deathKnight.FrostStrike)
+					} else if deathKnight.KillingMachineAura.IsActive() && deathKnight.RimeAura.IsActive() {
+						if deathKnight.CastCostPossible(sim, 0, 0, 1, 1) && deathKnight.CurrentRunicPower() < 110 {
+							*casted = deathKnight.DiseaseCheckWrapper(sim, target, deathKnight.HowlingBlast)
+						} else if deathKnight.CastCostPossible(sim, 0, 0, 1, 1) && deathKnight.CurrentRunicPower() > 110 {
+							*casted = deathKnight.DiseaseCheckWrapper(sim, target, deathKnight.HowlingBlast)
+						} else if !deathKnight.CastCostPossible(sim, 0, 0, 1, 1) && deathKnight.CurrentRunicPower() > 110 {
+							*casted = deathKnight.DiseaseCheckWrapper(sim, target, deathKnight.FrostStrike)
+						} else if !deathKnight.CastCostPossible(sim, 0, 0, 1, 1) && deathKnight.CurrentRunicPower() < 110 {
+							*casted = deathKnight.DiseaseCheckWrapper(sim, target, deathKnight.FrostStrike)
+						}
+					} else if !deathKnight.KillingMachineAura.IsActive() && deathKnight.RimeAura.IsActive() {
+						if deathKnight.CurrentRunicPower() < 110 {
+							*casted = deathKnight.DiseaseCheckWrapper(sim, target, deathKnight.HowlingBlast)
+						} else {
+							*casted = deathKnight.DiseaseCheckWrapper(sim, target, deathKnight.FrostStrike)
+						}
+					} else {
+						*casted = deathKnight.DiseaseCheckWrapper(sim, target, deathKnight.FrostStrike)
+						if !*casted {
+							*casted = deathKnight.DiseaseCheckWrapper(sim, target, deathKnight.HornOfWinter)
+						}
+					}
+				}
+			}
+		} else if deathKnight.opener.id == OpenerID_Unholy_Full {
+			// I suggest adding the a wrapper around each spell you cast like this:
+			// deathKnight.YourWrapper(sim, target, deathKnight.FrostStrike) that returns a bool for when you casted
+			// since the waiting code relies on knowing if you actually casted
+
 			if deathKnight.CanRaiseDead(sim) {
 				deathKnight.RaiseDead.Cast(sim, target)
-				return
+				*casted = true
+				return *casted
 			}
 			diseaseRefreshDuration := time.Duration(deathKnight.Rotation.DiseaseRefreshDuration) * time.Second
 			// Horn of Winter if you're the DK to refresh it and its not precasted/active
 			if deathKnight.ShouldHornOfWinter(sim) {
 				deathKnight.HornOfWinter.Cast(sim, target)
+				*casted = true
 			} else if (!deathKnight.TargetHasDisease(FrostFeverAuraLabel, target) || deathKnight.FrostFeverDisease[target.Index].RemainingDuration(sim) < diseaseRefreshDuration) && deathKnight.CanIcyTouch(sim) {
 				// Dont clip if theres half a second left to tick
 				remainingDuration := deathKnight.FrostFeverDisease[target.Index].RemainingDuration(sim)
@@ -64,6 +153,7 @@ func (deathKnight *DeathKnight) tryUseGCD(sim *core.Simulation) {
 					deathKnight.WaitUntil(sim, sim.CurrentTime+remainingDuration+1)
 				} else {
 					deathKnight.IcyTouch.Cast(sim, target)
+					*casted = true
 					recastedFF = true
 				}
 			} else if (!deathKnight.TargetHasDisease(BloodPlagueAuraLabel, target) || deathKnight.BloodPlagueDisease[target.Index].RemainingDuration(sim) < diseaseRefreshDuration) && deathKnight.CanPlagueStrike(sim) {
@@ -73,6 +163,7 @@ func (deathKnight *DeathKnight) tryUseGCD(sim *core.Simulation) {
 					deathKnight.WaitUntil(sim, sim.CurrentTime+remainingDuration+1)
 				} else {
 					deathKnight.PlagueStrike.Cast(sim, target)
+					*casted = true
 					recastedBP = true
 				}
 			} else {
@@ -84,35 +175,48 @@ func (deathKnight *DeathKnight) tryUseGCD(sim *core.Simulation) {
 					// Desolation and Pestilence check
 					if deathKnight.shouldSpreadDisease(sim) {
 						deathKnight.spreadDiseases(sim, target)
+						*casted = true
 					} else {
 						deathKnight.BloodStrike.Cast(sim, target)
+						*casted = true
 					}
 				} else {
 					if deathKnight.Rotation.UseDeathAndDecay {
 						// Death and Decay Rotation
 						if deathKnight.CanDeathAndDecay(sim) && deathKnight.AllDiseasesAreActive(target) {
 							deathKnight.DeathAndDecay.Cast(sim, target)
+							*casted = true
 						} else if deathKnight.CanGhoulFrenzy(sim) && deathKnight.Talents.MasterOfGhouls && (!deathKnight.Ghoul.GhoulFrenzyAura.IsActive() || deathKnight.Ghoul.GhoulFrenzyAura.RemainingDuration(sim) < 6*time.Second) && !deathKnight.shouldWaitForDnD(sim, false, false, true) {
 							deathKnight.GhoulFrenzy.Cast(sim, target)
+							*casted = true
 						} else if deathKnight.CanScourgeStrike(sim) && !deathKnight.shouldWaitForDnD(sim, false, true, true) {
 							deathKnight.ScourgeStrike.Cast(sim, target)
+							*casted = true
 						} else if !deathKnight.Talents.ScourgeStrike && deathKnight.CanIcyTouch(sim) && !deathKnight.shouldWaitForDnD(sim, false, true, false) {
 							deathKnight.IcyTouch.Cast(sim, target)
+							*casted = true
 						} else if !deathKnight.Talents.ScourgeStrike && deathKnight.CanPlagueStrike(sim) && !deathKnight.shouldWaitForDnD(sim, false, false, true) {
 							deathKnight.PlagueStrike.Cast(sim, target)
+							*casted = true
 						} else if deathKnight.CanBloodStrike(sim) && !deathKnight.shouldWaitForDnD(sim, true, false, false) {
 							if deathKnight.shouldSpreadDisease(sim) {
 								deathKnight.spreadDiseases(sim, target)
+								*casted = true
 							} else if deathKnight.Env.GetNumTargets() > 2 {
 								deathKnight.BloodBoil.Cast(sim, target)
+								*casted = true
 							} else {
 								deathKnight.BloodStrike.Cast(sim, target)
+								*casted = true
 							}
 						} else if deathKnight.CanDeathCoil(sim) && !deathKnight.SummonGargoyle.IsReady(sim) {
 							deathKnight.DeathCoil.Cast(sim, target)
+							*casted = true
 						} else if deathKnight.CanHornOfWinter(sim) {
 							deathKnight.HornOfWinter.Cast(sim, target)
+							*casted = true
 						} else {
+							// Probably want to make this just return *casted as casted should be false in this case, the wait time will be handled after the return
 							if deathKnight.GCD.IsReady(sim) && !deathKnight.IsWaiting() {
 								// This means we did absolutely nothing.
 								// Wait until our next auto attack to decide again.
@@ -127,20 +231,27 @@ func (deathKnight *DeathKnight) tryUseGCD(sim *core.Simulation) {
 						// Scourge Strike Rotation
 						if deathKnight.CanGhoulFrenzy(sim) && deathKnight.Talents.MasterOfGhouls && (!deathKnight.Ghoul.GhoulFrenzyAura.IsActive() || deathKnight.Ghoul.GhoulFrenzyAura.RemainingDuration(sim) < 6*time.Second) {
 							deathKnight.GhoulFrenzy.Cast(sim, target)
+							*casted = true
 						} else if deathKnight.CanScourgeStrike(sim) {
 							deathKnight.ScourgeStrike.Cast(sim, target)
+							*casted = true
 						} else if deathKnight.CanBloodStrike(sim) {
 							if deathKnight.shouldSpreadDisease(sim) {
 								deathKnight.spreadDiseases(sim, target)
+								*casted = true
 							} else if deathKnight.Env.GetNumTargets() > 2 {
 								deathKnight.BloodBoil.Cast(sim, target)
+								*casted = true
 							} else {
 								deathKnight.BloodStrike.Cast(sim, target)
+								*casted = true
 							}
 						} else if deathKnight.CanDeathCoil(sim) && !deathKnight.SummonGargoyle.IsReady(sim) {
 							deathKnight.DeathCoil.Cast(sim, target)
+							*casted = true
 						} else if deathKnight.CanHornOfWinter(sim) {
 							deathKnight.HornOfWinter.Cast(sim, target)
+							*casted = true
 						} else {
 							if deathKnight.GCD.IsReady(sim) && !deathKnight.IsWaiting() {
 								// This means we did absolutely nothing.
@@ -156,10 +267,29 @@ func (deathKnight *DeathKnight) tryUseGCD(sim *core.Simulation) {
 				}
 			}
 		}
-
-		// Start proper Frost rotation
-
-		// Frost DK rota
-		deathKnight.DoRotation(sim)
+		// Other prio lists for other specs here just else if {...
 	}
+
+	return *casted
+}
+
+func (deathKnight *DeathKnight) DoRotation(sim *core.Simulation) {
+	opener := deathKnight.opener
+	if !opener.DoNext(sim, deathKnight) {
+		if deathKnight.GCD.IsReady(sim) && !deathKnight.IsWaiting() {
+			waitUntil := deathKnight.AutoAttacks.MainhandSwingAt
+			if deathKnight.AutoAttacks.OffhandSwingAt > sim.CurrentTime {
+				waitUntil = core.MinDuration(waitUntil, deathKnight.AutoAttacks.OffhandSwingAt)
+			}
+			waitUntil = core.MinDuration(waitUntil, deathKnight.AnyRuneReadyAt(sim))
+			deathKnight.WaitUntil(sim, waitUntil)
+		} else { // No resources
+			waitUntil := deathKnight.AnySpentRuneReadyAt(sim)
+			deathKnight.WaitUntil(sim, waitUntil)
+		}
+	}
+}
+
+func (deathKnight *DeathKnight) ResetRotation(sim *core.Simulation) {
+	deathKnight.opener.idx = 0
 }
